Adde update and delete endpoints for products

+++ src/main/java/com/acme/rewear/platform/store/domain/model/commands/CreateProductCommand.java
@@ -1,5 +1,5 @@
 package com.acme.rewear.platform.store.domain.model.commands;
 
-public record CreateProductCommand(String name, String description, String size, String state, Float price, String color, String urlToImg) {
+public record CreateProductCommand(String name, String description, String size, String state, Double price, String color, String urlToImg) {
 
 }
