+++ src/main/java/com/acme/rewear/platform/store/interfaces/rest/resources/ProductResource.java
@@ -1,4 +1,4 @@
 package com.acme.rewear.platform.store.interfaces.rest.resources;
 
-public record ProductResource(Long id, String productName, String productDescription, String size, String state, Float price, String color, String urlToImg) {
+public record ProductResource(Long id, String productName, String productDescription, String size, String state, Double price, String color, String urlToImg) {
 }
